Fading Fog, QOL, crashfix
