+++ scripts/screens/event_screens.py
@@ -7,11 +7,12 @@ import pygame
 from scripts.events import events_class
 from scripts.utility import draw, get_text_box_theme, get_living_cat_count, scale
 # from scripts.game_structure.text import *
-from scripts.game_structure.image_button import IDImageButton
+from scripts.game_structure.image_button import IDImageButton, UIImageButton
 from scripts.game_structure.game_essentials import *
 from ..cat.cats import Cat
 from ..game_structure import image_cache
 from scripts.event_class import Single_Event
+from scripts.game_structure.windows import GameOver
 
 
 class EventsScreen(Screens):
